html/browsers/browsing-the-web/navigating-across-documents/010-form-action-javascript-url-document-writes-script-element.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/links.html#following-hyperlinks">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#queue-a-task">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#dom-manipulation-task-source">
<title>Link with onclick form submit to navigate "javascript: url" that document.writes a script  in a nested browsing context, is aborted by href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="010-form-action-javascript-url-document-writes-script-element-intermediary.html"></iframe>
<script>
var test = async_test();
window.onjsurl = window.onmessage = test.step_func_done(function(e) {
  assert_equals(e.data, "href", "'javascript: url' code is expected to run asynchronously.");
});
</script>

<!--
In an iframe...

1. This test creates an iframe, a form with `action="javascript: ..."` (which implicitly returns
    a string containing a script element that will attempt to call a synchronous handler in the intermediary) and an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will attempt to submit the form, which targets the iframe.
3. The form action "javascript: url" navigation is queued in a task which contains the
    steps to evaluate the code contents.
4. Navigating the iframe to the anchor's `href` must cancel the iframe
    "javascript: url" navigation.
5. If the semantics are implemented correctly, a single message will be
    received from href.html
    (There will be no message from the "javascript: url" code, because it
    is aborted by the navigation to "href.html")

-->
